WIP: Use tranformations in actions on node creation, as advised by Karsten
Sample configuraion:

```
'TYPO3.Neos.NodeTypes:Page':
options:
    actions:
      onCreate:
        -
          type: 'ChangePropertyValue'
          settings:
            property: 'title'
            newValue: '${q(node).property("title") + " - actionZZ"}'
```
